Refactoring
Added GridManager
GridGenerator destroyed
UndoManager destroyed Added "Q" to reset the scene
reason: Unable to implement the undo due to chaos reasons,
have an idea of how to implement
not sure why we need it

gameManager needs to be further refactored
Card manager due
